Migrate Realtime message tests
